+++ README.md
@@ -5,6 +5,6 @@ A version of Theo's [Zapdos](https://github.com/TheoBr/zapdos) app written with 
 - [ ] Stream embed + pin question
 - [ ] Delete individual questions + action to clear all questions
 - [ ] Optionally require auth for questions
-- [ ] Presence counter (?)
+- [x] Presence counter
 - [ ] style the login page
 - [ ] Add a footer with info & links
