+++ app/routes/auth/twitch.callback.tsx
@@ -1,5 +1,5 @@
 import type { LoaderFunction } from "@remix-run/node"
-import { authenticator } from "~/auth.server"
+import { authenticator } from "~/modules/core/auth.server"
 
 export const loader: LoaderFunction = ({ request }) =>
   authenticator.authenticate("twitch", request, { successRedirect: "/" })
