+++ app/routes/auth/twitch.login.tsx
@@ -1,6 +1,6 @@
 import type { ActionFunction, LoaderFunction } from "@remix-run/node"
 import { redirect } from "@remix-run/node"
-import { authenticator } from "~/auth.server"
+import { authenticator } from "~/modules/core/auth.server"
 
 export const loader: LoaderFunction = () => redirect("/")
