+++ app/routes/index.tsx
@@ -8,8 +8,8 @@ import { json } from "@remix-run/node"
 import { Form, useLoaderData } from "@remix-run/react"
 import { formatDistance, parseISO } from "date-fns"
 import { useState } from "react"
-import { authenticator } from "~/auth.server"
 import { useEventStream } from "~/helpers/event-stream"
+import { authenticator } from "~/modules/core/auth.server"
 import type { ClientQuestion } from "~/modules/questions/client-questions.server"
 import { loadClientQuestions } from "~/modules/questions/client-questions.server"
 import { useTimer } from "~/modules/react/use-timer"
