+++ package.json
@@ -37,6 +37,8 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "remix-auth": "^3.2.2",
+    "socket.io": "^4.5.1",
+    "socket.io-client": "^4.5.1",
     "zod": "^3.17.3"
   },
   "devDependencies": {
